Prepare a skeleton for Run command implementation
