+++ .gitignore
@@ -5,7 +5,6 @@
 /lefthook-local.yml
 /lefthook.yml
 /lefthook
-cplefthook
 
 tmp/
 dist/
