+++ cmd/lefthook/add-doc.txt
