+++ cmd/lefthook/main.go
@@ -3,13 +3,14 @@ package main
 import (
 	"os"
 
-	"github.com/evilmartians/lefthook/pkg/cmd"
+	"github.com/evilmartians/lefthook/internal/log"
 )
 
 func main() {
-	rootCmd := cmd.NewRootCmd()
+	rootCmd := newRootCmd()
 
 	if err := rootCmd.Execute(); err != nil {
+		log.Errorf("Error: %s", err)
 		os.Exit(1)
 	}
 }
