+++ internal/templates/config.tmpl
