+++ internal/templates/hook.tmpl
