--- pkg/version/version.go
@@ -1,7 +1,0 @@
-package version
-
-var version = "1.0.0"
-
-func Version() string {
-	return version
-}
